% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{optimization}
\alias{optimization}
\title{Linear Regression Via Gradient Descent Optimization}
\usage{
optimization(y, x, alpha = 0.05)
}
\arguments{
\item{y}{A \code{vector} of length n,whose entries are the values of the
response variable.}

\item{alpha}{A \code{number} that lies between \code{0} and \code{1}, which represent the confidence level. If not
specified, \code{alpha} is assigned a default value of 0.05.}

\item{X}{An n x p \code{matrix} or a \code{column vector} of length \code{n} whose columns represent the values of
p explanatory variables, and \code{n} rows are the observations.}
}
\value{
Returns a \code{list} containing the following attributes:
\describe{
     \item{betas}{A vector of estimated value of the coefficients}

     \item{CI}{Confidence intervals for the coefficients}

     \item{Rsq}{R-square value, which is a measure of goodness of fit of the model}

     \item{cp}{cp value, which is a measure of goodness of fit of the model}
     \item{F_stat}{F statistic for the model}

     \item{p-value}{p-value of the F_Stat}

     \item{Resid_plot}{Plot of residuals against fitted values}

     \item{qq-plot}{qq-plot of residulas}

     \item{hist_resid}{Histogram of residuals}
}
}
\description{
Performs linear regression via gradient descent optimization, and outputs the estimated coefficients,
confidence intervals for the coefficients, R square, cp, F statistics, and its p-value, plots of residual
against fitted values, qq-plot of residuals, and histogram of residuals.
}
\examples{
M<-matrix(rnorm(1000),ncol=10) #explanatory variable values
 y<-rnorm(100)               #response
optimization(y,M,0.01)

# Import data
abalone <- read.csv(file = 'abalone.csv')
# Transform dummy variables
library(fastDummies)
abalone <- dummy_cols(abalone, select_columns = "Sex")
# Create response variable and set of predictors
y <- abalone$Rings
x <- abalone[, 2:11]
x <- subset(x, select = -Rings)
optimization(y,x)
}
\author{
Carlos Matus, Ridvan Ozdemir, and Ogonnaya Romanus
}
